starting in on itemmanager and viewcontroller arch refactor
